Add createCreateHandler factory; share body schemas and row builders
POST create handlers for topics, tasks, providers, domains, routines and
dailies hand-rolled what createUpsertHandler already does for PUT. A new
createCreateHandler factory (uuid + insert + junction inserts + validate
+ afterCreate) replaces them, and each entity's body schema / row builder
/ junction builders now live in a single <entity>Rows.ts shared by both
handlers, so create and upsert can no longer drift apart.

Rows files import schemas relatively (not via @/) so node --test can
still load them directly.
