+++ packages/middleware/src/routes/api/dailies/createDaily.ts
@@ -1,73 +1,13 @@
-import { JsonSchemaToTsProvider } from "@fastify/type-provider-json-schema-to-ts";
-import { FastifyInstance } from "fastify";
-import { db } from "@/db";
 import { dailies } from "@/db/schema";
-import {
-  completionSchema,
-  criteriaSchema,
-  nullableDailyStatusEnum,
-  nullableString,
-} from "@/utils/schemas";
-import type { DailyCompletion, DailyCriteria } from "@emstack/types";
-import { v4 as uuidv4 } from "uuid";
+import { createCreateHandler } from "@/utils/createCreateHandler";
 
-const createSchema = {
-  schema: {
-    description: "Create a new daily",
-    body: {
-      type: "object",
-      required: ["name"],
-      properties: {
-        name: {
-          type: "string",
-        },
-        location: nullableString,
-        description: nullableString,
-        completions: {
-          type: "array",
-          items: completionSchema,
-        },
-        courseProviderId: nullableString,
-        resourceId: nullableString,
-        moduleGroupId: nullableString,
-        moduleId: nullableString,
-        taskId: nullableString,
-        status: nullableDailyStatusEnum,
-        criteria: criteriaSchema,
-      },
-    },
-  },
-} as const;
+import { buildDailyRow, dailyBodySchema } from "./dailyRows";
 
-export default async function (server: FastifyInstance) {
-  const fastify = server.withTypeProvider<JsonSchemaToTsProvider>();
+import type { DailyBody } from "./dailyRows";
 
-  fastify.post(
-    "/",
-    createSchema,
-    async function (request) {
-      const body = request.body;
-      const id = uuidv4();
-
-      await db.insert(dailies).values({
-        id,
-        name: body.name,
-        location: body.location ?? null,
-        description: body.description ?? null,
-        completions: (body.completions ?? []) as DailyCompletion[],
-        courseProviderId: body.courseProviderId ?? null,
-        resourceId: body.resourceId ?? null,
-        moduleGroupId: body.moduleGroupId ?? null,
-        moduleId: body.moduleId ?? null,
-        taskId: body.taskId || null,
-        status: body.status ?? "active",
-        criteria: (body.criteria ?? {}) as DailyCriteria,
-      });
-
-      return {
-        status: "ok",
-        id,
-      };
-    },
-  );
-}
+export default createCreateHandler<DailyBody>({
+  description: "Create a new daily",
+  table: dailies,
+  bodySchema: dailyBodySchema,
+  buildRow: buildDailyRow,
+});
