+++ packages/middleware/src/routes/api/providers/providerRows.ts
@@ -1,0 +1,58 @@
+import {
+  nullableBoolean,
+  nullableInteger,
+  nullableString,
+} from "../../../utils/schemas.ts";
+
+// Body schema and row builder shared by the provider create and upsert
+// handlers.
+
+export interface ProviderBody {
+  name: string;
+  url: string;
+  description?: string | null;
+  cost?: string | null;
+  isRecurring?: boolean | null;
+  recurDate?: string | null;
+  recurPeriodUnit?: "days" | "months" | "years" | null;
+  recurPeriod?: number | null;
+  isCourseFeesShared?: boolean | null;
+}
+
+export const providerBodySchema = {
+  type: "object",
+  required: ["name", "url"],
+  properties: {
+    name: {
+      type: "string",
+    },
+    description: nullableString,
+    url: {
+      type: "string",
+    },
+    cost: nullableString,
+    isRecurring: nullableBoolean,
+    recurDate: nullableString,
+    recurPeriodUnit: {
+      type: ["string", "null"],
+      enum: ["days", "months", "years", null],
+    },
+    recurPeriod: nullableInteger,
+    isCourseFeesShared: nullableBoolean,
+  },
+} as const;
+
+export function buildProviderRow(body: ProviderBody, id: string) {
+  return {
+    id,
+    name: body.name,
+    description: body.description ?? null,
+    url: body.url,
+    cost: body.cost ?? null,
+    isRecurring: body.isRecurring ?? null,
+    recurDate: body.recurDate ?? null,
+    recurPeriodUnit: body.recurPeriodUnit ?? undefined,
+    recurPeriod: body.recurPeriod ?? null,
+    isCourseFeesShared: body.isCourseFeesShared ?? null,
+  };
+}
