+++ packages/middleware/src/routes/api/routines/upsertRoutine.ts
@@ -1,63 +1,16 @@
 import { routineConnections, routines } from "@/db/schema";
 import { createUpsertHandler } from "@/utils/createUpsertHandler";
 import { buildRoutineConnectionRows } from "@/utils/routineConnectionRows";
-import {
-  completionSchema,
-  criteriaSchema,
-  nullableRoutineModeEnum,
-  nullableRoutineStatusEnum,
-  nullableString,
-  routineConnectionsSchema,
-  weeklySchema,
-} from "@/utils/schemas";
 
-import type { RoutineConnectionInput } from "@/utils/routineConnectionRows";
-import type { RoutineWeekly } from "@/db/schema";
-import type { DailyCompletion, DailyCriteria } from "@emstack/types";
+import { buildRoutineRow, routineBodySchema } from "./routineRows";
 
-interface RoutineBody {
-  name: string;
-  description?: string | null;
-  connections?: RoutineConnectionInput[];
-  status?: "active" | "inactive" | "complete" | "paused" | null;
-  weekly?: RoutineWeekly;
-  mode?: "weekly" | "daily" | null;
-  completions?: DailyCompletion[];
-  criteria?: DailyCriteria;
-}
+import type { RoutineBody } from "./routineRows";
 
 export default createUpsertHandler<RoutineBody>({
   description: "Create or update a routine",
   table: routines,
-  bodySchema: {
-    type: "object",
-    required: ["name"],
-    properties: {
-      name: {
-        type: "string",
-      },
-      description: nullableString,
-      connections: routineConnectionsSchema,
-      status: nullableRoutineStatusEnum,
-      weekly: weeklySchema,
-      mode: nullableRoutineModeEnum,
-      completions: {
-        type: "array",
-        items: completionSchema,
-      },
-      criteria: criteriaSchema,
-    },
-  },
-  buildRow: (body, id) => ({
-    id,
-    name: body.name,
-    description: body.description ?? null,
-    status: body.status ?? "active",
-    weekly: body.weekly ?? {},
-    mode: body.mode ?? "weekly",
-    completions: body.completions ?? [],
-    criteria: body.criteria ?? {},
-  }),
+  bodySchema: routineBodySchema,
+  buildRow: buildRoutineRow,
   updateableColumns: [
     "name",
     "description",
